Fix: No check added if course bundle is enabled

+++ ecommerce/HooksHandler.php
@@ -404,10 +404,12 @@ class HooksHandler {
 					 * For subscription, renewal no need to update order id.
 					 */
 					if ( $this->order_model->is_single_order( $order ) ) {
-						update_post_meta( $has_enrollment->ID, EnrollmentModel::ENROLLMENT_ORDER_ID_META, $order_id );
-						$has_bundle_enrollment_meta = get_post_meta( $has_enrollment->ID, CourseBundle::BUNDLE_ENROLLMENT_META, true );
-						if ( $has_bundle_enrollment_meta ) {
-							delete_post_meta( $has_enrollment->ID, CourseBundle::BUNDLE_ENROLLMENT_META );
+						if ( tutor_utils()->is_addon_enabled( 'course-bundle' ) ) {
+							update_post_meta( $has_enrollment->ID, EnrollmentModel::ENROLLMENT_ORDER_ID_META, $order_id );
+							$has_bundle_enrollment_meta = get_post_meta( $has_enrollment->ID, CourseBundle::BUNDLE_ENROLLMENT_META, true );
+							if ( $has_bundle_enrollment_meta ) {
+								delete_post_meta( $has_enrollment->ID, CourseBundle::BUNDLE_ENROLLMENT_META );
+							}
 						}
 						/**
 						 * Update enrollment expiry date if it is set in a course.
@@ -440,23 +442,21 @@ class HooksHandler {
 						do_action( 'tutor_order_enrolled', $order, $has_enrollment->ID );
 					}
 				}
-			} else {
-				if ( $order->order_status === $this->order_model::ORDER_COMPLETED ) {
+			} elseif ( $order->order_status === $this->order_model::ORDER_COMPLETED ) {
 					// Insert enrollment.
-					add_filter( 'tutor_enroll_data', fn( $enroll_data) => array_merge( $enroll_data, array( 'post_status' => 'completed' ) ) );
+					add_filter( 'tutor_enroll_data', fn( $enroll_data ) => array_merge( $enroll_data, array( 'post_status' => 'completed' ) ) );
 
 					$enrollment_id = EnrollmentModel::do_enroll( $object_id, $order_id, $student_id );
-					if ( $enrollment_id ) {
-						if ( $this->is_bundle_order( $order, $object_id ) && $this->order_model->is_single_order( $order ) ) {
-							BundleModel::enroll_to_bundle_courses( $object_id, $student_id );
-						}
-						update_post_meta( $enrollment_id, EnrollmentModel::ENROLLMENT_ORDER_ID_META, $order_id );
-
-						do_action( 'tutor_order_enrolled', $order, $enrollment_id );
-					} else {
-						// Log error message with student id and course id.
-						error_log( "Error updating enrollment for student {$student_id} and course {$object_id}" );
+				if ( $enrollment_id ) {
+					if ( $this->is_bundle_order( $order, $object_id ) && $this->order_model->is_single_order( $order ) ) {
+						BundleModel::enroll_to_bundle_courses( $object_id, $student_id );
 					}
+					update_post_meta( $enrollment_id, EnrollmentModel::ENROLLMENT_ORDER_ID_META, $order_id );
+
+					do_action( 'tutor_order_enrolled', $order, $enrollment_id );
+				} else {
+					// Log error message with student id and course id.
+					error_log( "Error updating enrollment for student {$student_id} and course {$object_id}" );
 				}
 			}
 		}
